Refine process pool
